Using the new listpack quicklist

+++ src/listpack.c
@@ -833,21 +833,21 @@ uint32_t lpBytes(unsigned char *lp) {
  * the tail, negative indexes specify elements starting from the tail, where
  * -1 means the last element, -2 the penultimate and so forth. If the index
  * is out of range, NULL is returned. */
-unsigned char *lpSeek(unsigned char *lp, long index) {
+unsigned char *lpSeek(unsigned char *lp, int index) {
     int forward = 1; /* Seek forward by default. */
 
     /* We want to seek from left to right or the other way around
      * depending on the listpack length and the element position.
      * However if the listpack length cannot be obtained in constant time,
      * we always seek from left to right. */
-    uint32_t numele = lpGetNumElements(lp);
+    int numele = lpGetNumElements(lp);
     if (numele != LP_HDR_NUMELE_UNKNOWN) {
-        if (index < 0) index = (long)numele+index;
+        if (index < 0) index = numele+index;
         if (index < 0) return NULL; /* Index still < 0 means out of range. */
-        if (index >= (long)numele) return NULL; /* Out of range the other side. */
+        if (index >= numele) return NULL; /* Out of range the other side. */
         /* We want to scan right-to-left if the element we are looking for
          * is past the half of the listpack. */
-        if (index > (long)numele/2) {
+        if (index > numele/2) {
             forward = 0;
             /* Right to left scanning always expects a negative index. Convert
              * our index to negative form. */
@@ -1040,7 +1040,7 @@ unsigned char *lpMerge(unsigned char **first, unsigned char **second) {
     return target;
 }
 
-unsigned char *lpDeleteRange(unsigned char *zl, long index, unsigned int num) {
+unsigned char *lpDeleteRange(unsigned char *zl, int index, unsigned int num) {
     unsigned int i = 0;
 
     unsigned char *p = lpSeek(zl,index);
